+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.8] -
+
+### [Changed]
+
+- Added `version --plain` and `plugins add --disregard/heed/ignore-warnings` flags
+  useful for scripting.
+
 ## [0.5.7] - 2024-01-22
 
 - Take latest version of encapsia-api